Installing of WebUI and minor fixes: - Adding `webui install` command to hv script - Create systemd service units for frontend/backend - Fix error handling in express.js for frontend
