+++ .github/workflows/build_release.yml
@@ -18,6 +18,11 @@ jobs:
         with:
           fetch-depth: 0  # Fetch all history for tags and commits
 
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '20'  # Adjust based on your Node.js version requirement
+
       - name: Extract tag and version info
         id: extract_info
         run: |
@@ -42,6 +47,60 @@ jobs:
             echo "RELEASE_NAME=[$BRANCH_NAME] Release $TAG_NAME" >> $GITHUB_OUTPUT
           fi
 
+      - name: Build Frontend
+        run: |
+          cd frontend
+          npm ci
+          npm run build
+          # The build script already creates dist/frontend, but we'll verify
+          ls -la dist/
+
+      - name: Build Backend
+        run: |
+          cd backend
+          npm ci
+          npm run build
+          # The build script already creates dist/backend, but we'll verify
+          ls -la dist/
+
+      - name: Create webui.zip
+        run: |
+          # Create a temporary directory for webui contents
+          mkdir -p webui_temp/frontend
+          mkdir -p webui_temp/backend
+          
+          # Copy frontend dist contents (everything inside dist/)
+          if [ -d "./frontend/dist" ]; then
+            cp -r ./frontend/dist/* ./webui_temp/frontend/
+          else
+            echo "Frontend dist directory not found"
+            exit 1
+          fi
+          
+          # Copy backend dist contents (everything inside dist/)
+          if [ -d "./backend/dist" ]; then
+            cp -r ./backend/dist/* ./webui_temp/backend/
+          else
+            echo "Backend dist directory not found"
+            exit 1
+          fi
+          
+          # Also copy backend's package.json and package-lock.json to backend folder
+          # (based on your backend build script, these are needed)
+          cp ./backend/package.json ./webui_temp/backend/
+          cp ./backend/package-lock.json ./webui_temp/backend/
+          
+          # Create the webui.zip
+          cd webui_temp
+          zip -r ../webui.zip .
+          cd ..
+          
+          # Clean up
+          rm -rf webui_temp
+          
+          # Verify the ZIP was created
+          ls -la webui.zip
+
       - name: Prepare setup files
         run: |
           # Create version.txt file
@@ -54,6 +113,9 @@ jobs:
         run: |
           # Create output directories
           mkdir -p output_zips/appdata
+          
+          # Move webui.zip to output directory
+          mv webui.zip output_zips/
           
           # Create appdata ZIP files
           if [ -d "./appdata" ]; then
@@ -162,3 +224,4 @@ jobs:
           files: |
             release_files/appdata/*.zip
             release_files/setup.zip
+            release_files/webui.zip
