+++ webui/backend/src/api/activity.ts
@@ -1,9 +1,13 @@
 import { AuthenticatedRequest } from "@/middleware/auth";
 import installer from "@/services/installer";
 import { CurrentActivity } from "@/types";
-import { Response } from "express";
+import { NextFunction, Response } from "express";
 
-export function getActivity(_req: AuthenticatedRequest, res: Response<CurrentActivity>) {
-    const activity = installer.getCurrentActivity() ?? { type: 'none' };
-    res.json(activity);
+export function getActivity(_req: AuthenticatedRequest, res: Response<CurrentActivity>, next: NextFunction) {
+    try {
+        const activity = installer.getCurrentActivity() ?? { type: "none" };
+        return res.json(activity);
+    } catch (error) {
+        next(error);
+    }
 }
