+++ webui/backend/src/api/modules.ts
@@ -1,13 +1,16 @@
 import { AuthenticatedRequest } from "@/middleware/auth";
-import { Response } from "express";
-import { logger } from "@/logger";
-import { ErrorResponse, GetModulesResponse } from "@/types";
+import { NextFunction, Response } from "express";
+import { GetModulesResponse } from "@/types";
 import installer from "@/services/installer";
 
-export async function getModules(_req: AuthenticatedRequest, res: Response<GetModulesResponse|ErrorResponse>) {
-    const response: GetModulesResponse = {
-        installedModules: await installer.getInstalledModules(),
-        availableModules: await installer.getAvailableModules(),
+export async function getModules(_req: AuthenticatedRequest, res: Response<GetModulesResponse>, next: NextFunction) {
+    try {
+        const response: GetModulesResponse = {
+            installedModules: await installer.getInstalledModules(),
+            availableModules: await installer.getAvailableModules(),
+        };
+        return res.json(response);
+    } catch (error) {
+        next(error);
     }
-    res.json(response);
 }
