+++ webui/backend/src/logger.ts
@@ -4,15 +4,8 @@ export const logger = createLogger({
     level: "info",
     format: format.json(),
     transports: [
-        new transports.File({ filename: "error.log", level: "error" }),
-        new transports.File({ filename: "combined.log" }),
-    ],
-});
-
-if (process.env.NODE_ENV !== "production") {
-    logger.add(
         new transports.Console({
             format: format.simple(),
         })
-    );
-}
+    ],
+});
